Library Reference

AlgebraicMetabolism.AlgebraicMetabolism — Module

Some description of ths package

AlgebraicMetabolism.SchMetabolicNet — Constant

ACSet definition for a Biochemical Systems Theory model

See Catlab.jl documentation for description of the @present syntax.

AlgebraicMetabolism.ReactionMetabolicNet — Type

ReactionMetabolicNet{R}

The main entry type for building a metabolic model with fixed parameters baked in.

AlgebraicMetabolism.dynamics_expr — Method

dynamics_expr(m::ReactionMetabolicNet)

Build the expression for a reaction net from the combinatorial data. The expression we want to build is equivalent to:

dxi = sum(μ[i,j] ⋅ γ[j] ⋅ prod(X[k]^f[j,k] for k in 1:N) for j in 1:N)

$\frac{d}{dt} X_i = \sum_j(\mu_{i,j} \cdot \gamma_j \cdot \prod_k X_k^f_{j,k}$

This formula evaluates the dynamics of the system.

AlgebraicMetabolism.edges₁ — Method

edges₁(m::ReactionMetabolicNet, i::Int, j::Int)

access a vector of the E₁ edges between vertex i and vertex j.

AlgebraicMetabolism.edges₂ — Method

edges₂(m::ReactionMetabolicNet, i::Int, j::Int)

access a vector of the E₂ edges between vertex i and vertex j.

AlgebraicMetabolism.SSystems — Module

Some description of ths package

AlgebraicMetabolism.SSystems.System — Type

System{R}

The main entry type for building a metabolic model with fixed parameters baked in.

AlgebraicMetabolism.SSystems.edges₁ — Method

edges₁(m::AbstractMetabolicNet, i::Int, j::Int)

access a vector of the E₁ edges between vertex i and vertex j.

AlgebraicMetabolism.SSystems.edges₂ — Method

edges₂(m::AbstractMetabolicNet, i::Int, j::Int)

access a vector of the E₂ edges between vertex i and vertex j.

AlgebraicMetabolism.dynamics_expr — Method

dynamics_expr(m::System)

Build the expression for an S-System from the combinatorial data. The expression we want to build is equivalent to:

$\frac{d}{dt} X_i = \alpha_i \prod_j X_j^{g_{i,j}} - \beta_i\prod_j X^{h_{i,j}}$

This formula evaluates the dynamics of the system.